+++ expo-app/src/data/welcome.json
@@ -14,13 +14,13 @@
   {
     "page": 3,
     "title": "Geração de Roteiro com IA",
-    "description": "A aplicação utiliza a OpenAI para gerar roteiros únicos e adaptados às suas preferências de viagem",
+    "description": "Conheça Felipe! Ele utiliza a OpenAI para gerar roteiros únicos e adaptados às suas preferências de viagem",
     "image": "Ai.svg"
   },
   {
     "page": 4,
     "title": "Converse com seu Guia Turístico",
-    "description": "Você agora possui um Guia Turístico funcionando com Inteligência Artificial para te ajudar no que fazer a qualquer momento!",
+    "description": "Felipe será seu Guia Turístico funcionando com Inteligência Artificial para te ajudar no que fazer a qualquer momento!",
     "image": "AiChat.svg"
   },
   {
