+++ expo-app/src/screens/AIChat/AIChat.tsx
