+++ expo-app/src/screens/AIChat/AIChatMenu.tsx
@@ -1,0 +1,7 @@
+import { View } from "@gluestack-ui/themed";
+
+export function AIChatMenu(){
+  return(
+    <View></View>
+  )
+}
