+++ expo-app/src/screens/Itinerary/GenerateItinerary.tsx
